Move rs485_posix into serial_posix. Rename Enable to Enabled.
